requirements.txt
# GhostStream - Transcoding Service Dependencies
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Configuration
pyyaml>=6.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.0

# mDNS/Zeroconf
zeroconf>=0.131.0

# HTTP Client (for fetching source files)
httpx>=0.27.0
aiofiles>=23.2.1

# WebSocket support (included in uvicorn[standard])
websockets>=12.0

# Job Queue
asyncio-throttle>=1.0.2

# Logging
python-json-logger>=2.0.7

# Utilities
psutil>=5.9.7

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
